[#4390] - refine implementation and add function documentation
